usize it all!

+++ palmettobug/ImageProcessing/ImageAnalysisClass.py
@@ -930,8 +930,8 @@ class ImageAnalysis:
                     output = rsc.mask_boolean_rust(np.ascontiguousarray(mask1, dtype=np.uint), 
                                                    np.ascontiguousarray(mask2, dtype=np.uint), 
                                                    kind = kind, 
-                                                   object_threshold = object_threshold, 
-                                                   pixel_threshold = pixel_threshold, 
+                                                   object_threshold = np.uint(object_threshold), 
+                                                   pixel_threshold = np.uint(pixel_threshold), 
                                                    re_order = True)
                 else:
                     output = self._mask_bool(mask1, mask2, kind = kind, object_threshold = object_threshold, pixel_threshold = pixel_threshold)
